chore: extract layoutprovider
